Move fetch text UI route and tests to ui_fragments module
